requirements.txt
# core
Django==2.1.7
djangorestframework==3.9.1
mysql-connector-python==8.0.13
mysqlclient==1.3.14
jsonschema==2.6.0
netaddr==0.7.19
responses==0.10.5
jsonrpcserver==4.0.1

# web
gunicorn==19.9.0

# tests
coverage==4.5.2
pytest==4.1.1
pytest-django==3.4.5
pytest-cov==2.6.1
mixer==6.1.3


